cloudwalk-chatbot/requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers>=2.7.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
